requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
Pillow==10.4.0
tensorflow==2.16.1
